restructure volmesh and add aliases
