+++ src/compas/datastructures/halfface/__init__.py
@@ -1,0 +1,1 @@
+from .halfface import HalfFace  # noqa: F401
