+++ src/compas/datastructures/volmesh/__init__.py
@@ -2,11 +2,12 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from .core import *  # noqa: F401 F403
-from ._volmesh import *  # noqa: F401 F403
+from .operations import *  # noqa: F401 F403
 
 from .bbox import *  # noqa: F401 F403
 from .transformations import *  # noqa: F401 F403
+
+from .volmesh import *  # noqa: F401 F403
 
 
 __all__ = [name for name in dir() if not name.startswith('_')]
