--- src/compas/datastructures/volmesh/core/__init__.py
@@ -1,8 +1,0 @@
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-from .halfface import HalfFace  # noqa: F401
-from .volmesh import BaseVolMesh  # noqa: F401
-
-__all__ = [name for name in dir() if not name.startswith('_')]
